% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi_available_templates.R
\name{gi_available_templates}
\alias{gi_available_templates}
\title{Fetch available templates from gitignore.io}
\usage{
gi_available_templates()
}
\value{
A character with all templates supported by gitignore.io.
}
\description{
This return list of all templates supported by gitignore.io.
}
\details{
The returned list is returned as lower case characters.
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
gi_available_templates()
\dontshow{\}) # examplesIf}
}
