% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitignore_write_gitignore.R
\name{gi_write_gitignore}
\alias{gi_write_gitignore}
\title{Append or create a .gitignore file}
\usage{
gi_write_gitignore(new_lines, gitignore_file = here::here(".gitignore"))
}
\arguments{
\item{new_lines}{Template returned by `gi_fetch_templates()`.}

\item{gitignore_file}{Path of the .gitignore file to modify.}
}
\description{
Use the returned template(s) to append the existing .gitignore file.
}
