% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitignore_fetch_available_templates.R
\name{gitignore_fetch_available_templates}
\alias{gitignore_fetch_available_templates}
\title{Fetch available templates from gitignore.io}
\usage{
gitignore_fetch_available_templates()
}
\value{
A character vector with all operating systems, IDE or programming languages supported by gitignore.io.
}
\description{
This return list of all operating systems, IDE or programming languages supported by gitignore.io.
}
\details{
The returned list is returned as lower case characters.
}
\examples{
gitignore_fetch_available_templates()
}
