% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitignore_fetch_ignore_templates.R
\name{gitignore_fetch_ignore_templates}
\alias{gitignore_fetch_ignore_templates}
\title{Fetch the git ignore template from gitignore.io}
\usage{
gitignore_fetch_ignore_templates(template_name, copy_to_clipboard = TRUE)
}
\arguments{
\item{template_name}{A character vector with values included in `gitignore_fetch_available_templates()`.}

\item{copy_to_clipboard}{Logical, should the returned template(s) be copied to the clipboard? Otherwise, it will be printed in the console.}
}
\value{
A gitignore template.
}
\description{
Fetch the git ignore template from gitignore.io
}
\examples{
# Fetch template for the R language
gitignore_fetch_ignore_templates("R")

# You can combine many templates at once
gitignore_fetch_ignore_templates(c("R", "python", "java"))
}
