% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitignore_write_gitignore.R
\name{gitignore_write_gitignore}
\alias{gitignore_write_gitignore}
\title{Append .gitignore file}
\usage{
gitignore_write_gitignore(new_lines)
}
\arguments{
\item{new_lines}{Template returned by `gitignore_fetch_ignore_templates()`.}
}
\description{
Use returned template(s) to append the existing .gitignore file.
}
