% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitignore_fetch_ignore_templates.R
\name{gitignore_fetch_ignore_templates}
\alias{gitignore_fetch_ignore_templates}
\title{Fetch the git ignore template from gitignore.io}
\usage{
gitignore_fetch_ignore_templates(template_name)
}
\arguments{
\item{template_name}{A character vector with values included in `gitignore_fetch_available_templates()`.}
}
\value{
A gitignore template.
}
\description{
Fetch the git ignore template from gitignore.io
}
